verify: DD-37-003 — Add PresenceUpdate variant to WsServerMessage and fix StatusReport field types

+++ services/data-broker/src/ws.rs
@@ -188,7 +188,7 @@ async fn handle_socket(mut socket: WebSocket, state: AppState, user_id: Uuid) {
                                 let cfg = &state.config;
                                 let new_level = compute_throttle(
                                     current,
-                                    render_fps,
+                                    render_fps as f64,
                                     pending_updates,
                                     cfg.throttle_fps_low,
                                     cfg.throttle_fps_high,
